+++ NEWS.md
@@ -1,6 +1,7 @@
 
 # eatGADS 0.16.0.9000
 
+* `compareGADS()` now works for character variables
 * `subImputations()` for substituting imputations with unimputed values
 * `applyChangeMeta()` now returns a more informative error if the variable sets in the `GADSdat` and the `changeTable` differ
 * `dummies2char()` for recoding a set of dummy variables to a set of left filled character variables